Fix: Refactor code, remove unnecessary logs and use the dynamic url from the env.
