+++ client/package.json
@@ -24,6 +24,9 @@
   },
   "devDependencies": {
     "@eslint/js": "^9.13.0",
+    "@testing-library/jest-dom": "^6.6.3",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "^14.5.2",
     "@types/axios": "^0.14.4",
     "@types/node": "^22.10.0",
     "@types/react": "^18.3.12",
@@ -38,8 +41,12 @@
     "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-react-refresh": "^0.4.14",
     "globals": "^15.11.0",
+    "jest": "^29.7.0",
+    "msw": "^2.6.6",
     "postcss": "^8.4.49",
+    "redux-mock-store": "^1.5.5",
     "tailwindcss": "^3.4.15",
+    "ts-jest": "^29.2.5",
     "typescript": "~5.6.2",
     "vite": "^5.4.10",
     "vitest": "^2.1.5"
