+++ client/src/App.tsx
@@ -2,8 +2,8 @@
 import React, { useEffect } from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
-import { fetchUsers } from "./redux/reducers/userSlice";
-import { fetchAlbums } from "./redux/reducers/albumSlice";
+import { fetchUsers } from "./redux/slices/userSlice";
+import { fetchAlbums } from "./redux/slices/albumSlice";
 import LandingPage from "./components/LandingPage";
 import Home from "./components/Home";
 import UserDetails from "./components/UserDetails";
