+++ client/src/components/Header.tsx
@@ -2,8 +2,8 @@ import React, { useState, useEffect } from "react";
 import { FaBars, FaTimes, FaUserCircle, FaAngleLeft } from "react-icons/fa";
 import { useDispatch, useSelector } from "react-redux";
 import { RootState } from "../redux/store";
-import { logoutUser, fetchAuthUser } from "../redux/reducers/authSlice";
-import { AppDispatch } from "../redux/store"; // Import AppDispatch type
+import { logoutUser, fetchAuthUser } from "../redux/slices/authSlice
+import { AppDispatch } from "../redux/store";
 
 const Header: React.FC = () => {
   const [isNavOpen, setIsNavOpen] = useState<boolean>(false);
