+++ client/src/redux/slices/authSlice.ts
@@ -18,18 +18,23 @@ const initialState: AuthState = {
   id: undefined
 };
 
+const devUrl = process.env.DEV_URL;
+const prodUrl = process.env.PROD_URL;
+
+const baseUrl =
+  process.env.NODE_ENV === "production"
+    ? prodUrl
+    : devUrl;
+
 // Async thunk for logging out
 export const logoutUser = createAsyncThunk(
   "auth/logoutUser",
   async (_, { rejectWithValue }) => {
     try {
-      console.log("Logging out..."); // Log before sending logout request
-      await axios.get("http://localhost:5000/api/auth/logout", {
+      await axios.get(`${baseUrl}/api/auth/logout`, {
         withCredentials: true,
       });
-      console.log("Logged out successfully"); // Log if logout is successful
     } catch (error: any) {
-      console.error("Error logging out:", error); // Log any error that occurs
       return rejectWithValue(error.response?.data || "Error logging out");
     }
   },
@@ -40,14 +45,11 @@ export const fetchAuthUser = createAsyncThunk(
   "auth/fetchAuthUser",
   async (_, { rejectWithValue }) => {
     try {
-      console.log("Fetching user details..."); // Log before sending request
-      const response = await axios.get("http://localhost:5000/api/auth/me", {
+      const response = await axios.get(`${baseUrl}/api/auth/me`, {
         withCredentials: true,
       });
-      console.log("Fetched user details:", response.data); // Log successful response
       return response.data;
     } catch (error: any) {
-      console.error("Error fetching user details:", error); // Log any error
       return rejectWithValue(
         error.response?.data || "Error fetching user details",
       );
@@ -61,7 +63,6 @@ const authSlice = createSlice({
   reducers: {
     // Normal logout without an API call
     logout(state) {
-      console.log("Logging out (normal logout without API call)"); // Log normal logout
       state.email = null;
       state.isAuthenticated = false;
     },
@@ -69,36 +70,30 @@ const authSlice = createSlice({
   extraReducers: (builder) => {
     builder
       .addCase(fetchAuthUser.pending, (state) => {
-        console.log("Fetching user details... pending..."); // Log before state is updated to loading
         state.loading = true;
         state.error = null;
       })
       .addCase(fetchAuthUser.fulfilled, (state, action) => {
-        console.log("User details fetched successfully:", action.payload);
         state.loading = false;
-        state.id = action.payload.id; // Set userId correctly
+        state.id = action.payload.id;
         state.email = action.payload.email;
         state.isAuthenticated = true;
       })
       .addCase(fetchAuthUser.rejected, (state, action) => {
-        console.log("Failed to fetch user details:", action.payload); // Log error state update
         state.loading = false;
         state.error = action.payload as string;
         state.isAuthenticated = false;
       })
       .addCase(logoutUser.pending, (state) => {
-        console.log("Logout in progress... pending..."); // Log before state is updated to loading
         state.loading = true;
         state.error = null;
       })
       .addCase(logoutUser.fulfilled, (state) => {
-        console.log("User logged out successfully"); // Log successful logout state update
         state.loading = false;
         state.email = null;
         state.isAuthenticated = false;
       })
       .addCase(logoutUser.rejected, (state, action) => {
-        console.log("Failed to log out:", action.payload); // Log error state update
         state.loading = false;
         state.error = action.payload as string;
       });
